// ts/src/generated/protos/akash/gov/v1beta3/params_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts,json_types=true"
// @generated from file akash/gov/v1beta3/params.proto (package akash.gov.v1beta3, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import { file_gogoproto_gogo } from "../../../gogoproto/gogo_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file akash/gov/v1beta3/params.proto.
 */
export const file_akash_gov_v1beta3_params: GenFile = /*@__PURE__*/
  fileDesc("Ch5ha2FzaC9nb3YvdjFiZXRhMy9wYXJhbXMucHJvdG8SEWFrYXNoLmdvdi52MWJldGEzIrsBCg1EZXBvc2l0UGFyYW1zEqkBChhtaW5faW5pdGlhbF9kZXBvc2l0X3JhdGUYASABKAxChgHI3h8A2t4fJmdpdGh1Yi5jb20vY29zbW9zL2Nvc21vcy1zZGsvdHlwZXMuRGVj4t4fFU1pbkluaXRpYWxEZXBvc2l0UmF0ZereHxhtaW5faW5pdGlhbF9kZXBvc2l0X3JhdGXy3h8feWFtbDoibWluX2luaXRpYWxfZGVwb3NpdF9yYXRlIkIhWh9wa2cuYWt0LmRldi9nby9ub2RlL2dvdi92MWJldGEzYgZwcm90bzM", [file_gogoproto_gogo]);

/**
 * DepositParams defines the parameters for the x/gov module.
 *
 * @generated from message akash.gov.v1beta3.DepositParams
 */
export type DepositParams = Message<"akash.gov.v1beta3.DepositParams"> & {
  /**
   * MinInitialDepositRate is the minimum % of TotalDeposit
   * author of the proposal must put in order for proposal tx to be committed.
   *
   * @generated from field: bytes min_initial_deposit_rate = 1;
   */
  minInitialDepositRate: Uint8Array;
};

/**
 * DepositParams defines the parameters for the x/gov module.
 *
 * @generated from message akash.gov.v1beta3.DepositParams
 */
export type DepositParamsJson = {
  /**
   * MinInitialDepositRate is the minimum % of TotalDeposit
   * author of the proposal must put in order for proposal tx to be committed.
   *
   * @generated from field: bytes min_initial_deposit_rate = 1;
   */
  minInitialDepositRate?: string;
};

/**
 * Describes the message akash.gov.v1beta3.DepositParams.
 * Use `create(DepositParamsSchema)` to create a new message.
 */
export const DepositParamsSchema: GenMessage<DepositParams, DepositParamsJson> = /*@__PURE__*/
  messageDesc(file_akash_gov_v1beta3_params, 0);