// ts/src/generated/protos/akash/gov/v1beta3/genesis_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts,json_types=true"
// @generated from file akash/gov/v1beta3/genesis.proto (package akash.gov.v1beta3, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import { file_gogoproto_gogo } from "../../../gogoproto/gogo_pb";
import type { DepositParams, DepositParamsJson } from "./params_pb";
import { file_akash_gov_v1beta3_params } from "./params_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file akash/gov/v1beta3/genesis.proto.
 */
export const file_akash_gov_v1beta3_genesis: GenFile = /*@__PURE__*/
  fileDesc("Ch9ha2FzaC9nb3YvdjFiZXRhMy9nZW5lc2lzLnByb3RvEhFha2FzaC5nb3YudjFiZXRhMyJ5CgxHZW5lc2lzU3RhdGUSaQoOZGVwb3NpdF9wYXJhbXMYASABKAsyIC5ha2FzaC5nb3YudjFiZXRhMy5EZXBvc2l0UGFyYW1zQi/I3h8A6t4fDmRlcG9zaXRfcGFyYW1z8t4fFXlhbWw6ImRlcG9zaXRfcGFyYW1zIkIhWh9wa2cuYWt0LmRldi9nby9ub2RlL2dvdi92MWJldGEzYgZwcm90bzM", [file_gogoproto_gogo, file_akash_gov_v1beta3_params]);

/**
 * GenesisState stores slice of genesis deployment instance.
 *
 * @generated from message akash.gov.v1beta3.GenesisState
 */
export type GenesisState = Message<"akash.gov.v1beta3.GenesisState"> & {
  /**
   * DepositParams defines all the parameters related to deposits.
   *
   * @generated from field: akash.gov.v1beta3.DepositParams deposit_params = 1;
   */
  depositParams?: DepositParams;
};

/**
 * GenesisState stores slice of genesis deployment instance.
 *
 * @generated from message akash.gov.v1beta3.GenesisState
 */
export type GenesisStateJson = {
  /**
   * DepositParams defines all the parameters related to deposits.
   *
   * @generated from field: akash.gov.v1beta3.DepositParams deposit_params = 1;
   */
  depositParams?: DepositParamsJson;
};

/**
 * Describes the message akash.gov.v1beta3.GenesisState.
 * Use `create(GenesisStateSchema)` to create a new message.
 */
export const GenesisStateSchema: GenMessage<GenesisState, GenesisStateJson> = /*@__PURE__*/
  messageDesc(file_akash_gov_v1beta3_genesis, 0);